% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.R
\name{data_prep}
\alias{data_prep}
\title{Preparing data}
\usage{
data_prep(
  comp = NULL,
  vars_for_grouping = NULL,
  bin_lab = "bin",
  count_var = "count"
)
}
\arguments{
\item{comp}{A compositional dataset with columns for counts (can be names differently) per bin (can be named differently), and all `vars_for_grouping`}

\item{vars_for_grouping}{A character vector of variables used for grouping the compositional dataframe}

\item{bin_lab}{A string with the column label for the bins if not "bin"}

\item{count_lab}{A string with the column label for the counts by bins if not "counts"}
}
\description{
This function takes a dataset of counts (number of samples) per bin
(length/age/etc) across a range of strata (vars_for_grouping) and expands
it for use in the poisson-multinomial model
}
