% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{Fx_plot}
\alias{Fx_plot}
\title{Effects}
\usage{
Fx_plot(comp_data, mod, grp = "Year", form = "~(1|bin:Year)", grid = NULL)
}
\arguments{
\item{comp_data}{Compositional dataframe}

\item{mod}{An object from fit_model}

\item{grp}{group to do PPC over}

\item{form}{formula for effect}
}
\description{
Effects
}
