% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{post_pred_group}
\alias{post_pred_group}
\title{Post predictive}
\usage{
post_pred_group(mod, grp = NULL, xlab = "Length (cm)")
}
\arguments{
\item{mod}{An object from fit_model}

\item{grp}{group to do PPC over}

\item{xlab}{label for composition plot}
}
\description{
Post predictive
}
