% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{scale_comps}
\alias{scale_comps}
\title{scale compositions}
\usage{
scale_comps(
  scale_df,
  predvar = "catch",
  fit = NULL,
  grps,
  iters = NULL,
  form = NULL,
  pgrid = NULL
)
}
\arguments{
\item{scale_df}{data frame used for scaling}

\item{predvar}{Column in the scale_df used for scaling compositions}

\item{fit}{An object from fit_model}

\item{grps}{groups to do scaling}

\item{form}{an alternative random effect structure to the fitted model}
}
\description{
scale compositions
}
