% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc.R
\name{scaled_comp_plot}
\alias{scaled_comp_plot}
\title{Plot scaled compositions}
\usage{
scaled_comp_plot(
  scaled_comp = NULL,
  grps = NULL,
  scaled = T,
  scales = "free_y",
  comp_are = NULL,
  comp_are2 = NULL,
  cvar = "tcatch"
)
}
\arguments{
\item{scaled_comp}{data frame used for scaling}

\item{grps}{groups to do scaling}

\item{scaled}{proportions or catch-at}

\item{scales}{free y axis?}
}
\description{
Plot scaled compositions
}
