% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Selectivity}
\alias{Selectivity}
\title{Selectivity}
\usage{
Selectivity(bins, L50, L95)
}
\arguments{
\item{bins}{Bin structure}

\item{L50}{50 pc selectivity}

\item{l95}{95 pc selectivity offset from L50}
}
\description{
Selectivity
}
