% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.R
\name{fit_model}
\alias{fit_model}
\title{Fit the poisson multinomial model}
\usage{
fit_model(
  form,
  data = NULL,
  dist = "poisson",
  chains = 4,
  threads = 16,
  cores = 32,
  backend = "cmdstanr",
  iter = 1000,
  warmup = 500,
  thin = 2,
  refresh = 10
)
}
\arguments{
\item{form}{A RHS formula for the poisson-multinomial model, without explicit mention of bins}
}
\description{
Fit the poisson multinomial model
}
