% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.R
\name{parse_form}
\alias{parse_form}
\title{Parse a formula}
\usage{
parse_form(form = "gear + area + area:yy")
}
\arguments{
\item{form}{A RHS formula for the poisson-multinomial model, without explicit mention of bins}
}
\description{
Appends a formula by attaching a bin interaction term
}
